Add files via upload
The goal is to use a crank-slider mechanism to provide the motion according to the desired motion curve. It's a kind of robot inverse kinematic problem in which we have the position of end-effector, and we should obtain the orientation of joints. Here, we have the motion curve of the slider, and the angular position, velocity and acceleration of the crank should be computed.